Fix two regressions
- Wrong socket used to send UDP reply
- Synchronization issues in the UDP client maps

Introduced in 1c2c900365c4947ffd5e1f3821e8b133a961cd52.
1. The client map structure does not check for the socket used to
receive messages when determining whether a client exists. When the
wrong socket is used to send a reply, client `recv` will fail.

2. When a client is added to the maps, the client ID map is locked
first, then the client address map is locked, which leaves a possibility
that another client with the same address seeing the old address map,
determining that the client doesn't exist, and adding a new entry to the
maps. Later, when the first client is pruned from the maps, the second
client is also pruned. There is also a possibility of a deadlock
(fixed), which have not been observed in practice.

Signed-off-by: Zhang Maiyun <me@myzhangll.xyz>