Fix v6.0 read block regression
Two ends may deadlock waiting for each other to Ack.
Note that `tokio::select!` only processes the branches concurrently,
and once one branch returns, other ones are canceled while the body
of the `select!` is being executed. This means that if we are trying to
write when `frame_rx` receives an `Ack`, it will never be processed.

We cannot use `FutureExt::then` either because it will get cancelled
and cause a loss of data.

Signed-off-by: Zhang Maiyun <me@maiyun.me>